+++ default_www/install/engine/step_4.php
@@ -44,7 +44,7 @@ class InstallerStep4 extends InstallerStep
 	 */
 	public static function isAllowed()
 	{
-		return (isset($_SESSION['db_hostname']) && isset($_SESSION['db_database']) && isset($_SESSION['db_username']) && isset($_SESSION['db_password']));
+		return (isset($_SESSION['default_language']) && isset($_SESSION['default_interface_language']) && isset($_SESSION['multiple_languages']) && isset($_SESSION['languages']) && isset($_SESSION['interface_languages']));
 	}
 
 
@@ -78,32 +78,6 @@ class InstallerStep4 extends InstallerStep
 
 		// add multi checkbox
 		$this->frm->addMultiCheckbox('modules', $modules, array_unique(array_merge($this->modules['required'], $checkedModules)));
-
-		// seperate frontend/backend languages?
-		$this->frm->addCheckbox('same_interface_language', (SpoonSession::exists('same_interface_language') ? SpoonSession::get('same_interface_language') : true));
-
-		// multiple or single language (frontend)
-		$this->frm->addRadiobutton('language_type',	array(array('value' => 'multiple', 'label' => 'Multiple languages', 'variables' => array('multiple' => true)),
-													array('value' => 'single', 'label' => 'Just one language', 'variables' => array('single' => true))), (SpoonSession::exists('multiple_languages') && SpoonSession::get('multiple_languages')) ? 'multiple' : 'single');
-
-		// multiple languages (frontend)
-		$this->frm->addMultiCheckbox('languages', array(array('value' => 'en', 'label' => 'English'),
-														array('value' => 'fr', 'label' => 'French'),
-														array('value' => 'nl', 'label' => 'Dutch')), (SpoonSession::exists('languages') ? SpoonSession::get('languages') : 'nl'));
-
-		// multiple languages (backend)
-		$this->frm->addMultiCheckbox('interface_languages', array(array('value' => 'en', 'label' => 'English'),
-																	array('value' => 'fr', 'label' => 'French'),
-																	array('value' => 'nl', 'label' => 'Dutch')), (SpoonSession::exists('interface_languages') ? SpoonSession::get('interface_languages') : 'nl'));
-
-		// single language (frontend)
-		$this->frm->addDropdown('language', array('en' => 'English', 'fr' => 'French', 'nl' => 'Dutch'), (SpoonSession::exists('default_language') ? SpoonSession::get('default_language') : 'en'));
-
-		// default language (frontend)
-		$this->frm->addDropdown('default_language', array('en' => 'English', 'fr' => 'French', 'nl' => 'Dutch'), (SpoonSession::exists('default_language') ? SpoonSession::get('default_language') : 'en'));
-
-		// default language (backend)
-		$this->frm->addDropdown('default_interface_language', array('en' => 'English', 'fr' => 'French', 'nl' => 'Dutch'), (SpoonSession::exists('default_interface_language') ? SpoonSession::get('default_interface_language') : 'en'));
 
 		// example data
 		$this->frm->addCheckbox('example_data', (SpoonSession::exists('example_data') ? SpoonSession::get('example_data') : true));
@@ -143,40 +117,6 @@ class InstallerStep4 extends InstallerStep
 		// form submitted
 		if($this->frm->isSubmitted())
 		{
-			// multiple languages
-			if($this->frm->getField('language_type')->getValue() == 'multiple')
-			{
-				// list of languages
-				$languages = $this->frm->getField('languages')->getValue();
-
-				// default language
-				if(!in_array($this->frm->getField('default_language')->getValue(), $languages)) $this->frm->getField('default_language')->setError('Your default language needs to be in the list of languages you chose.');
-			}
-
-			// single language
-			else
-			{
-				// list of languages
-				$languages = (array) array($this->frm->getField('default_language')->getValue());
-			}
-
-			// same cms interface language
-			if($this->frm->getField('same_interface_language')->getChecked())
-			{
-				// list of languages
-				$interfaceLanguages = $languages;
-			}
-
-			// different interface language
-			else
-			{
-				// list of languages
-				$interfaceLanguages = $this->frm->getField('interface_languages')->getValue();
-			}
-
-			// default language
-			if(!in_array($this->frm->getField('default_interface_language')->getValue(), $interfaceLanguages)) $this->frm->getField('default_interface_language')->setError('Your default language needs to be in the list of languages you chose.');
-
 			// all valid
 			if($this->frm->isCorrect())
 			{
@@ -189,12 +129,7 @@ class InstallerStep4 extends InstallerStep
 				// set modules
 				SpoonSession::set('modules', $modules);
 
-				// set languages
-				SpoonSession::set('default_language', $this->frm->getField('default_language')->getValue());
-				SpoonSession::set('default_interface_language', $this->frm->getField('default_interface_language')->getValue());
-				SpoonSession::set('multiple_languages', ($this->frm->getField('language_type')->getValue() == 'multiple') ? true : false);
-				SpoonSession::set('languages', $languages);
-				SpoonSession::set('interfaceLanguages', $interfaceLanguages);
+				// example data
 				SpoonSession::set('example_data', $this->frm->getField('example_data')->getChecked());
 
 				// redirect